requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
amqp==5.1.1
aniso8601==9.0.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
base58==2.1.1
billiard==3.6.4.0
bitarray==2.6.0
blinker==1.5
Brotli==1.0.9
CacheControl==0.12.11
cachelib==0.9.0
cachetools==5.2.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
commonmark==0.9.1
cryptography==39.0.1
cytoolz==0.12.0
Deprecated==1.2.13
distlib==0.3.6
eth-hash==0.3.3
eth-typing==3.2.0
eth-utils==2.0.0
Faker==15.0.0
filelock==3.8.0
firebase-admin==5.1.0
Flask==2.1.3
Flask-And-Redis==1.0.0
Flask-JWT-Extended==4.4.4
Flask-Limiter==2.6.3
flask-marshmallow==0.14.0
Flask-PyMongo==2.3.0
flask-restx==0.5.1
frozenlist==1.3.1
gevent==21.12.0
google-api-core==2.10.0
google-api-python-client==2.60.0
google-auth==2.14.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
greenlet==1.1.3.post0
hexbytes==0.3.0
hyper==0.7.0
httplib2==0.21.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipfshttpclient==0.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
kombu==5.2.4
limits==2.7.0
MarkupSafe==2.1.1
marshmallow==3.18.0
msgpack==1.0.4
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
oauth2client==4.1.3
packaging==21.3
parsimonious==0.10.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
prompt-toolkit==3.0.31
proto-plus==1.22.1
protobuf==4.21.9
py3rijndael==0.3.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pyOpenSSL==23.0.0
pycryptodome==3.17
pycparser==2.21
pycryptodome==3.17
pydash==5.1.1
Pygments==2.13.0
pyhumps==3.7.3
PyJWT==2.5.0
pymongo==4.2.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-http-client==3.3.7
python-slugify==6.1.2
pytz==2022.4
redis==4.3.4
regex==2022.9.13
requests==2.28.1
rich==12.6.0
rlp==3.0.0
rsa==4.9
secure-smtplib==0.1.1
semantic-version==2.10.0
sendgrid==6.9.7
sentry-sdk==1.7.0
six==1.16.0
starkbank-ecdsa==2.1.0
text-unidecode==1.3
toml==0.10.2
toolz==0.12.0
typing_extensions==4.4.0
ua-parser==0.16.1
uritemplate==4.1.1
urllib3==1.26.12
uWSGI==2.0.19.1
varint==1.0.2
vine==5.0.0
wcwidth==0.2.5
websockets==10.3
Werkzeug==2.1.2
wrapt==1.14.1
yarl==1.8.1
zipp==3.9.0
zope.event==4.5.0
zope.interface==5.5.0
